fix(portal): modernize async runtime and enable service at install
Replace manual event loop with asyncio.run() in portal backend, add
--break-system-packages for Arch pip compatibility, enable portal
service on install. Also fix stale yazi-frontend references in PRD,
add --any-display to CLAUDE.md IPC example, add signed/unsigned cast
in filesystemmodel, and clean up minor QML comments.

+++ plugin/src/Symmetria/FileManager/Models/filesystemmodel.cpp
@@ -156,7 +156,7 @@ int FileSystemModel::rowCount(const QModelIndex& parent) const {
 }
 
 QVariant FileSystemModel::data(const QModelIndex& index, int role) const {
-    if (role != Qt::UserRole || !index.isValid() || index.row() >= m_entries.size()) {
+    if (role != Qt::UserRole || !index.isValid() || index.row() >= static_cast<int>(m_entries.size())) {
         return QVariant();
     }
     return QVariant::fromValue(m_entries.at(index.row()));
